requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
black==26.1.0
cfgv==3.5.0
click==8.3.1
coverage==7.13.4
distlib==0.4.0
exceptiongroup==1.3.1
fastapi==0.129.0
filelock==3.24.1
flake8==7.3.0
h11==0.16.0
identify==2.6.16
idna==3.11
iniconfig==2.3.0
librt==0.8.0
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
pdf2image==1.17.0
pillow==12.1.1
platformdirs==4.9.1
pluggy==1.6.0
pre_commit==4.5.1
pycodestyle==2.14.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytesseract==0.3.13
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
python-magic==0.4.27
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
SQLAlchemy==2.0.46
starlette==0.52.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
virtualenv==20.36.1
